refactor test_andscript_code
