+++ test/price_utils.js
@@ -2,7 +2,7 @@ const BN = require("bn.js")
 const assert = require("assert")
 const Contract = require("@truffle/contract")
 
-const { addCustomMintableTokenToExchange } = require("./test_utils")
+const { addCustomMintableTokenToExchange } = require("../scripts/utils/test_and_script_code")(web3, artifacts)
 const {
   isPriceReasonable,
   checkNoProfitableOffer,
