+++ test/verification_script.js
@@ -8,7 +8,9 @@ const ProxyFactory = artifacts.require("GnosisSafeProxyFactory")
 const EvilGnosisSafeProxy = artifacts.require("EvilGnosisSafeProxy")
 
 const { verifyCorrectSetup } = require("../scripts/utils/verify_scripts")(web3, artifacts)
-const { addCustomMintableTokenToExchange, createTokenAndGetData, deploySafe, populatePriceStorage } = require("./test_utils")
+const { addCustomMintableTokenToExchange, deploySafe } = require("../scripts/utils/test_and_script_code")(web3, artifacts)
+
+const { createTokenAndGetData, populatePriceStorage } = require("./test_utils")
 const { execTransaction, waitForNSeconds } = require("../scripts/utils/internals")(web3, artifacts)
 const {
   getAllowances,
